requirements.txt
click==8.1.3
Flask==2.2.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
Werkzeug==2.2.3
mysql-connector-python~=8.0.5
grpcio==1.54.0
grpcio-tools==1.54.0

